Add Discount Grpc

+++ src/Services/Basket/Basket.API/Models/ShoppingCartItem.cs
@@ -6,7 +6,7 @@ public class ShoppingCartItem
 
     public string Color { get; set; } = default!;
 
-    public decimal Price { get; set; } = default!;
+    public double Price { get; set; } = default!;
 
     public Guid ProductId { get; set; } = default!;
